+++ android/app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.view.LoginActivity"
+            android:name=".presentation.view.login.LoginActivity"
             android:exported="true"
             android:theme="@style/Theme.Chongdae">
             <intent-filter>
